Add option to form Floquet Hamiltonian
This updates cis_Floquet.py and simulation.py so that one can choose to
form the full H_fl and use sparse LU factorization routines
(scipy.sparse.linalg.spilu or MKL_PARDISO for now) or use Block-LU
factorization. Input file structure is updated accordingly.
Also updated Floquet_system.H_linop to use the shift in the matvec.
Added a check of the accuracy of the solvers.